+++ CHANGELOG.md
@@ -22,6 +22,14 @@
 - `c-admin-display`: Binary values now render as links that will download the value as a file, instead of only showing the length in bytes.
 - `c-datetime-picker`: Added prop `showTodayButton`
 
+# 5.3.8
+
+- Fix error in bulk saves where inheritance workaround added in 5.3.6 was not specific enough.
+
+# 5.3.7
+
+- Fix error in bulk saves where children of a model that has been `$remove()`'d where neither the children nor the removed model have PKs would still attempt to create.
+
 # 5.3.6
 
 - Fix error in bulk saves where models that belong to inheritance hierarchies would get skipped over while performing collection fixup.